requirements.txt
pyserial>=3.5
RPi.GPIO>=0.7.0
# Note: BlynkLib is included as a local file, but modern blynk-library-python can also be used.
